# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: ComfyUI Serverless API
  version: 1.0.0
  description: |
    Serverless ComfyUI API for AI image generation and workflow execution.
    
    This API provides a complete solution for running ComfyUI workflows in a serverless environment,
    with features including:
    - Asynchronous workflow execution
    - Model management and caching  
    - File upload/download
    - Real-time progress tracking
    - Webhook notifications
    - Authentication and rate limiting
    
  contact:
    name: API Support
    email: support@comfyui-serverless.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.comfyui-serverless.com
    description: Production server
  - url: https://staging-api.comfyui-serverless.com
    description: Staging server
  - url: http://localhost:8000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /models/:
    get:
      summary: List Models
      description: List all available models with their status
      tags:
        - Models
      parameters:
        - name: model_type
          in: query
          schema:
            $ref: '#/components/schemas/ModelType'
          description: Filter by model type
        - name: available_only
          in: query
          schema:
            type: boolean
            default: false
          description: Show only available models
      responses:
        '200':
          description: List of models
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelListResponse'
  /models/{model_name}:
    get:
      summary: Get Model Status
      description: Get detailed status for a specific model
      tags:
        - Models
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelStatus'
        '404':
          description: Model not found
  /models/{model_name}/download:
    post:
      summary: Download Model
      description: Download a model from URL
      tags:
        - Models
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
        - name: model_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ModelType'
        - name: download_url
          in: query
          required: true
          schema:
            type: string
            format: uri
        - name: description
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Download started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  model_name:
                    type: string
                  model_type:
                    type: string
  /models/{model_name}/load:
    post:
      summary: Load Model
      description: Load model into memory
      tags:
        - Models
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
        - name: model_type
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ModelType'
      responses:
        '200':
          description: Model loaded successfully
  /models/{model_name}/unload:
    post:
      summary: Unload Model
      description: Unload model from memory
      tags:
        - Models
      parameters:
        - name: model_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Model unloaded successfully
  /models/cleanup:
    post:
      summary: Cleanup Unused Models
      description: Clean up models that haven't been used recently
      tags:
        - Models
      parameters:
        - name: max_age_hours
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 24
            default: 1
      responses:
        '200':
          description: Cleanup completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  models_unloaded:
                    type: array
                    items:
                      type: string

  # File Management endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ModelType:
      type: string
      enum:
        - checkpoint
        - lora
        - embedding
        - vae
        - controlnet
        - upscaler
      example: "checkpoint"
    ModelStatus:
      type: object
      required:
        - name
        - type
        - is_loaded
        - is_downloading
      properties:
        name:
          type: string
          example: "sd-v1-5-inpainting.ckpt"
        type:
          $ref: '#/components/schemas/ModelType'
        is_loaded:
          type: boolean
        is_downloading:
          type: boolean
        download_progress:
          type: number
          format: float
          minimum: 0.0
          maximum: 1.0
          nullable: true
        last_used:
          type: string
          format: date-time
          nullable: true
        memory_usage_mb:
          type: number
          format: float
          nullable: true
    ModelListResponse:
      type: object
      required:
        - models
        - total_memory_usage_mb
        - available_memory_mb
      properties:
        models:
          type: array
          items:
            $ref: '#/components/schemas/ModelStatus'
        total_memory_usage_mb:
          type: number
          format: float
        available_memory_mb:
          type: number
          format: float

    # File schemas
...